🚚 Rename: 파일 위치 변경 및 경로 수정
