+++ src/components/recruitPage/Question.jsx
@@ -1,11 +1,11 @@
 import { useState } from "react";
-import styles from "../../styles/recruitPage/question.module.css";
-import arrowIcon from "../../assets/recruitPage/down.svg";
+import styles from "./question.module.css";
+import arrowIcon from "@/assets/recruitPage/down.svg";
 
 export default function Question() {
     const qaList = [
         {
-            question: "비전공자도 참여 능한가요?", 
+            question: "비전공자도 참여 가가능한가요?", 
             answer:
                 "비전공자도 참여 가능합니다.\n멋쟁이사자처럼의 목표는 컴퓨터과학 비전공자들도 프로그래밍 기초 지식을 배워 자신의 웹서비스를 만들 수 있도록 하는 것입니다.",
         },
