+++ src/components/resultPage/checkResult.module.css
