+++ src/components/resultPage/passPage.module.css
